Classes and methods (#4)
* cleanup of structure and old files

* added flag to delete collection

* trying abstract methods in ollamaclient

* new structure

* needed for new structure

* fixing naming convention

* combined unstructured and refactored

* refactored more imports

* cleaned up prompt

* added chunk display

* switched to system prompting

* tried cleaning

* added progress bars

* adding ragas notebook

* switched to False

+++ holodeck/rag/pipeline.py
@@ -1,23 +1,25 @@
 from weaviate.collections import Collection
-import holodeck.rag.ollama_utils as ollama_utils
-import holodeck.rag.weaviate_utils as weaviate_utils
+import holodeck.ollama.embeddings as embeddings
+from holodeck.ollama.ollama_client import OllamaClient
+import holodeck.weaviate.weaviate_utils as weaviate_utils
 import holodeck.utilities.constants as constants
 from loguru import logger
 from unstructured.staging.base import convert_to_dict
 
-def pipeline_prep() -> None:
+def pipeline_prep(delete_collection: bool) -> None:
     
     logger.info("Getting embeddings client...")
-    embeddingClient = ollama_utils.get_embeddings_client(constants.OLLAMA_LOCAL_URL)
+    embeddingClient = OllamaClient.get_client(host=constants.OLLAMA_LOCAL_URL)
     logger.info("Setting up embedding model...")
-    embeddingModel = ollama_utils.setup_embedding_model(embeddingClient, constants.DEFAULT_EMBEDDING_MODEL)    
+    embeddingModel = OllamaClient.setup_model(embeddingClient, model=constants.DEFAULT_EMBEDDING_MODEL)    
     
     logger.info("Creating Weaviate client...")
     weaviateClient = weaviate_utils.create_weaviate_local_client()
     
     # adding this to delete basic chunking
-    logger.info("Deleting Weaviate collection...")
-    weaviate_utils.delete_collection(weaviateClient, constants.WEAVIATE_COLLECTION_NAME)
+    if delete_collection:
+        logger.info("Deleting Weaviate collection...")
+        weaviate_utils.delete_collection(weaviateClient, constants.WEAVIATE_COLLECTION_NAME)
     
     logger.info("Getting Weaviate collection...")
     weaviateCollection = weaviate_utils.get_collection(weaviateClient, constants.WEAVIATE_COLLECTION_NAME)
@@ -28,5 +30,5 @@ def pipeline_prep() -> None:
     else:
         logger.info("Elements found to add to Weaviate collection")
         elementDictionary = convert_to_dict(elements=elements)
-        elementChunks = ollama_utils.generate_embeddings(embeddingModel, elementDictionary)
+        elementChunks = embeddings.generate_embeddings(embeddingModel, elementDictionary)
         weaviate_utils.load_chunks_into_weaviate(elementChunks, weaviateClient, weaviateCollection)
